+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 
 - Avoid choosing mutations that aren't applicable in the current context.
 - Response schema validation is now 5–15x faster on average.
+- Response header schema validation is now 15–30x faster on average.
 
 ## [4.1.4](https://github.com/schemathesis/schemathesis/compare/v4.1.3...v4.1.4) - 2025-09-01